fix: use the new decoding also in the wallet home
